+++ NEWS.md
@@ -5,6 +5,7 @@
 + New functions for algorithmic trading at the experimental stage.
 
   + `tradeStrategy()` is an example of a Moving Average crossover strategy.
+  + `futuresMonths` data set for futures contract month codes.
 
 ## Enhancement